exporting the compact theme
